// src/search/interfaces/repoCard/irepo.card.view.controller.ts
import React from 'react';
import {ResultSearchGithubRepositoryEntity} from '../../bussiness/entities/result.search.github.repository.entity';

export interface IRepoCardViewController {
  repository: ResultSearchGithubRepositoryEntity;

  switchState: boolean;
  setSwitchState: React.Dispatch<React.SetStateAction<boolean>>;

  loading: boolean;
  setLoading: React.Dispatch<React.SetStateAction<boolean>>;

  context: any;

  saveRepositoryToSniffer: () => Promise<void>;

  removeSnifferRepository: () => Promise<void>;
}